+++ src/Storage/Storage.Management/ChangeLog.md
@@ -33,6 +33,9 @@
 * Supported create/update storage blob container with RootSquash
     - `New-AzRmStorageContainer`
     - `Update-AzRmStorageContainer`
+* Supported AllowProtectedAppendWriteAll in set container Immutability Policy, and add container LegalHold
+    - `Set-AzRmStorageContainerImmutabilityPolicy`
+    - `Add-AzRmStorageContainerLegalHold`
 
 ## Version 3.11.0
 * Supported get/set blob tags on a specific blob